rename: Button을 TextButton으로 수정, pages에 페이지 컴포넌트 이동
