+++ src/pages/CardDone/index.jsx
@@ -1,17 +1,13 @@
 import React from "react";
 import PropTypes from "prop-types";
 
-import ButtonContainer from "../../containers/ButtonContainer";
+import TextButton from "../../components/TextButton";
 
 const CardDone = ({ process, setProcess }) => {
 	return (
 		<div className="app">
 			<h2>카드 추가완료 페이지</h2>
-			<ButtonContainer
-				process={process}
-				setProcess={setProcess}
-				content="확인"
-			/>
+			<TextButton process={process} setProcess={setProcess} content="확인" />
 		</div>
 	);
 };
