+++ src/pages/CardList/index.jsx
@@ -1,13 +1,13 @@
 import React from "react";
 import PropTypes from "prop-types";
 
-import ButtonContainer from "../../containers/ButtonContainer";
+import TextButton from "../../components/TextButton";
 
 const CardList = ({ process, setProcess }) => {
 	return (
 		<div className="app">
 			<h2>카드 목록 페이지</h2>
-			<ButtonContainer process={process} setProcess={setProcess} />
+			<TextButton process={process} setProcess={setProcess} />
 		</div>
 	);
 };
